changed the hierarquy of structure and added a test with Mars

+++ laws.py
@@ -3,7 +3,7 @@ class LawsOfPhysics:
         pass
     @classmethod
     def gravitationalEnergyEq(self, mass, gravity, height):
-        return (mass * gravity * height)
+        return float(mass * gravity * height)
     @classmethod
     def kineticEnergyEq(self, mass, vel):
-        return ( mass * (vel**2) ) / 2
+        return float( (mass * (vel**2)) / 2 )
